Added a lot
I manually added the ocamlgraph library so that way we do not need to worry about getting it.
Added a make file so now all you should have to do is run "make" in the terminal to get an output.
Added an OCamlMakefile to go along with the make file.
Added a testProgram.exe runnable so everyone can see what the output looks like now.