Classification addition and name changes
Created the classification ml file. It is commented and completed unless additions need to be made later.
Made name changes to the Decision tree file, and got rid of the first version of the file to declutter a little.
Fixed the naming in the python script that Noah made so it should continue to work.
Updated the Makefile to include the Str package, have the name changes, include the classifer, and now create the parsedot test program.